Fix incorrect type of numerator and denominator fields of Coefficients
These were still erroneously Float64s rather than StationXML.Floats.

+++ src/derived_types.jl
@@ -211,8 +211,8 @@ $(DocStringExtensions.TYPEDFIELDS)
 @with_kw mutable struct Coefficients
     @BaseFilter
     cf_transfer_function_type::CfTransferFunction
-    numerator::Vector{Float64} = Float64[]
-    denominator::Vector{Float64} = Float64[]
+    numerator::Vector{Float} = Float[]
+    denominator::Vector{Float} = Float[]
 end
 
 attribute_fields(::Type{Coefficients}) = BASE_FILTER_ATTRIBUTES
